# contract/openapi.json
{
  "components" : {
    "schemas" : {
...
      "MembershipResponseDTO" : {
        "properties" : {
          "joinedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "membershipId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "personEmail" : {
            "type" : "string"
          },
          "personId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "personName" : {
            "type" : "string"
          },
          "professionalCredential" : {
            "type" : "string"
          },
          "role" : {
            "enum" : [ "VETERINARIO", "MONITOR", "VOLUNTARIO", "ADMINISTRADOR" ],
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "MembershipRoleRequestDTO" : {
        "properties" : {
          "role" : {
            "enum" : [ "VETERINARIO", "MONITOR", "VOLUNTARIO", "ADMINISTRADOR" ],
            "type" : "string"
          }
        },
        "required" : [ "role" ],
        "type" : "object"
      },
...
      "OrganizationRequestDTO" : {
        "properties" : {
          "address" : {
            "type" : "string"
          },
          "cep" : {
            "type" : "string"
          },
          "city" : {
            "type" : "string"
          },
          "cnpj" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "ownerVetName" : {
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "state" : {
            "type" : "string"
          }
        },
        "required" : [ "name" ],
        "type" : "object"
      },
      "OrganizationResponseDTO" : {
        "properties" : {
          "address" : {
            "type" : "string"
          },
          "cep" : {
            "type" : "string"
          },
          "city" : {
            "type" : "string"
          },
          "cnpj" : {
            "type" : "string"
          },
          "creationDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "organizationId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "ownerVetName" : {
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "state" : {
            "type" : "string"
          },
          "updateDate" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "bearerFormat" : "JWT",
        "description" : "O token devolvido por POST /auth/login.",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.0.1",
  "paths" : {
...
    "/organizations/members/{membershipId}" : {
      "delete" : {
        "description" : "Marca a saida e nao apaga o vinculo - o que ela registrou continua no historico dos animais. Idempotente.",
        "operationId" : "removeMember",
        "parameters" : [ {
          "in" : "path",
          "name" : "membershipId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "summary" : "Desliga a pessoa da equipe",
        "tags" : [ "organization-member-controller" ]
      },
      "patch" : {
        "description" : "So administrador. Rebaixar o ultimo administrador e recusado: sem ele ninguem convida, ajusta nem desliga.",
        "operationId" : "changeMemberRole",
        "parameters" : [ {
          "in" : "path",
          "name" : "membershipId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MembershipRoleRequestDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MembershipResponseDTO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Ajusta a funcao de quem e da equipe",
        "tags" : [ "organization-member-controller" ]
      }
    },
    "/organizations/{organizationId}" : {
      "delete" : {
        "operationId" : "deleteOrganization",
        "parameters" : [ {
          "in" : "path",
          "name" : "organizationId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "summary" : "Apaga a organizacao",
        "tags" : [ "organization-controller" ]
      },
      "get" : {
        "operationId" : "getOrganizationById",
        "parameters" : [ {
          "in" : "path",
          "name" : "organizationId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrganizationResponseDTO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Uma organizacao",
        "tags" : [ "organization-controller" ]
      },
      "put" : {
        "description" : "Atualizacao parcial: apenas os campos presentes no payload sao alterados. Campos ausentes ou nulos preservam o valor existente. Nenhum campo e obrigatorio neste endpoint.",
        "operationId" : "updateOrganization",
        "parameters" : [ {
          "in" : "path",
          "name" : "organizationId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OrganizationRequestDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrganizationResponseDTO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Atualiza dados da clinica",
        "tags" : [ "organization-controller" ]
      }
    },
...
  },
  "security" : [ {
    "bearer-jwt" : [ ]
  } ]
}